rename more references
